requirements.txt
flask
flask-cors
langchain
langchain-community
tiktoken
pyyaml
atlassian-python-api
python-docx
gunicorn
